Notes varint frontend

+++ backend/src/controllers/notes.ts
@@ -2,10 +2,15 @@ import NoteModel from "../models/note";
 import {RequestHandler} from "express";
 import createHttpError from "http-errors";
 import mongoose from "mongoose";
+import {assertIsDefined} from "../util/assertIsDefined";
 
 export const getNotes: RequestHandler = async (req, res, next) => {
+    const authenticatedUserId = await req.session.userId;
+
     try {
-        const notes = await NoteModel.find().exec();
+        assertIsDefined(authenticatedUserId);
+
+        const notes = await NoteModel.find({userId: authenticatedUserId}).exec();
         res.status(200).json(notes);
     } catch (error) {
         next(error);
@@ -14,8 +19,11 @@ export const getNotes: RequestHandler = async (req, res, next) => {
 
 export const getNote:RequestHandler = async (req, res, next) => {
     const noteid = req.params.noteid;
+    const authenticatedUserId = await req.session.userId;
 
     try {
+        assertIsDefined(authenticatedUserId);
+
         if(!mongoose.isValidObjectId(noteid)) {
             throw createHttpError(400, "Not a valid Note Id");
         }
@@ -24,6 +32,10 @@ export const getNote:RequestHandler = async (req, res, next) => {
 
         if(!note){
             throw createHttpError(404, "Note not found");
+        }
+
+        if(!note.userId.equals(authenticatedUserId)) {
+            throw createHttpError(401, "You cannot access this note");
         }
 
         res.status(200).json(note);
@@ -40,14 +52,18 @@ interface CreateNodeBody{
 export const createNote:RequestHandler<unknown, unknown, CreateNodeBody, unknown> = async (req, res, next) => {
     const title = req.body.title;
     const text = req.body.text;
+    const authenticatedUserId = await req.session.userId;
 
     try {
+        assertIsDefined(authenticatedUserId);
+
         if(!title){
             throw createHttpError(400, "Title is required"  );
         }
 
 
         const newNote = await NoteModel.create({
+            userId: authenticatedUserId,
             title: title,
             text: text,
         })
@@ -72,8 +88,11 @@ export const updateNote:RequestHandler<UpdateNoteParams, unknown, UpdateNoteBody
     const noteid = req.params.noteid;
     const newTitle = req.body.title;
     const newText = req.body.text;
+    const authenticatedUserId = await req.session.userId;
 
     try {
+        assertIsDefined(authenticatedUserId);
+
         if(!mongoose.isValidObjectId(noteid)) {
             throw createHttpError(400, "Not a valid Note Id");
         }
@@ -87,6 +106,10 @@ export const updateNote:RequestHandler<UpdateNoteParams, unknown, UpdateNoteBody
         if(!note){
             throw createHttpError(404, "Note not found");
         }
+        if(!note.userId.equals(authenticatedUserId)) {
+            throw createHttpError(401, "You cannot access this note");
+        }
+
 
         note.title = newTitle;
         note.text = newText;
@@ -103,7 +126,10 @@ export const updateNote:RequestHandler<UpdateNoteParams, unknown, UpdateNoteBody
 
 export const deleteNote:RequestHandler = async(req,res,next) => {
     const noteid = req.params.noteid;
+    const authenticatedUserId = await req.session.userId;
+
     try {
+        assertIsDefined(authenticatedUserId);
 
         if(!mongoose.isValidObjectId(noteid)) {
             throw createHttpError(400, "Not a valid Note Id");
@@ -114,6 +140,11 @@ export const deleteNote:RequestHandler = async(req,res,next) => {
         if(!note){
             throw createHttpError(404, "Note not found");
         }
+
+        if(!note.userId.equals(authenticatedUserId)) {
+            throw createHttpError(401, "You cannot access this note");
+        }
+
 
         await note.deleteOne();
